requirements.txt
python-telegram-bot==13.7
python-dotenv==0.19.0
APScheduler==3.9.1
pytz==2021.1 